(fga-eps-mds/2023-1-CAPJu-Doc#114) - Conserta arquitetura e maus cheiros de codigo
